+++ test/testChainr.js
@@ -1,7 +1,7 @@
-define(['parse/parse', 'parse/parse_lang', 'parse/parse_string'],
+define(['parse/parse', 'parse/lang', 'parse/string'],
 function(parse, parse_lang, parse_string){
-    var add = parse.next(parse.character('+'), parse.always(function(x, y){ return x + y; }));
-    var mul = parse.next(parse.character('*'), parse.always(function(x, y){ return x * y; }));
+    var add = parse.next(parse_string.character('+'), parse.always(function(x, y){ return x + y; }));
+    var mul = parse.next(parse_string.character('*'), parse.always(function(x, y){ return x * y; }));
     var op = parse.either(add, mul);
     
     var num = parse.bind(parse_string.digit, function(x){ return parse.always(parseInt(x)); });