+++ test/testEndBy.js
@@ -1,10 +1,10 @@
-define(['parse/parse', 'parse/parse_lang'], function(parse, parse_lang){
+define(['parse/parse', 'parse/string', 'parse/lang'], function(parse, parse_string, parse_lang){
     return {
         'module': "EndBy",
         'tests': [
             ["Simple endBy",
             function(){
-                var a = parse.eager(parse_lang.endBy(parse.character(','), parse.character('a')));
+                var a = parse.eager(parse_lang.endBy(parse_string.character(','), parse_string.character('a')));
                 
                 assert.deepEqual(parse.run(a, "a,a,"), ['a', 'a']);
                 
@@ -14,7 +14,7 @@ define(['parse/parse', 'parse/parse_lang'], function(parse, parse_lang){
             }],
             ["Simple endBy1",
             function(){
-                var a = parse.eager(parse_lang.endBy1(parse.character(','), parse.character('a')));
+                var a = parse.eager(parse_lang.endBy1(parse_string.character(','), parse_string.character('a')));
                 
                 assert.deepEqual(parse.run(a, "a,a,"), ['a', 'a']);