+++ test/testSepBy.js
@@ -1,10 +1,10 @@
-define(['parse/parse', 'parse/parse_lang'], function(parse, parse_lang){
+define(['parse/parse', 'parse/string', 'parse/lang'], function(parse, parse_string, parse_lang){
     return {
-        'module': "SepBy Tests",
+        'module': "parse_lang.sepBy",
         'tests': [
             ["Simple sepBy",
             function(){
-                var a = parse.eager(parse_lang.sepBy(parse.character(','), parse.character('a')));
+                var a = parse.eager(parse_lang.sepBy(parse_string.character(','), parse_string.character('a')));
                 
                 assert.deepEqual(parse.run(a, "a,a"), ['a', 'a']);
                 
@@ -14,16 +14,16 @@ define(['parse/parse', 'parse/parse_lang'], function(parse, parse_lang){
             }],
             ["No backtracking",
             function(){
-                var a = parse.eager(parse_lang.sepBy(parse.character(','), parse.character('a')));
+                var a = parse.eager(parse_lang.sepBy(parse_string.character(','), parse_string.character('a')));
                 
-                assert.throws(parse.run.bind(undefined, a, "a,a,z"), parse.UnexpectError);
+                assert.throws(parse.run.bind(undefined, a, "a,a,z"), parse.ExpectError);
                 
                 assert.deepEqual(parse.run(a, ","), []);
             }],
             
             ["Simple sepBy1",
             function(){
-                var a = parse.eager(parse_lang.sepBy1(parse.character(','), parse.character('a')));
+                var a = parse.eager(parse_lang.sepBy1(parse_string.character(','), parse_string.character('a')));
                 
                 assert.deepEqual(parse.run(a, "a,a"), ['a', 'a']);