+++ test/testThen.js
@@ -1,15 +1,15 @@
-define(['parse/parse', 'parse/parse_lang'], function(parse, parse_lang){
+define(['parse/parse', 'parse/string', 'parse/lang'], function(parse, parse_string, parse_lang){
     return {
-        'module': "Then Tests",
+        'module': "parse_lang.then",
         'tests': [
             ["Simple Then",
             function(){
-                var a = parse_lang.then(parse.character('a'), parse.character('b'));
+                var a = parse_lang.then(parse_string.character('a'), parse_string.character('b'));
                 
                 assert.deepEqual(parse.run(a, "ab"), 'a');
                 
                 assert.deepEqual(
-                    parse.run(parse.next(a, parse.character('c')), "abc"),
+                    parse.run(parse.next(a, parse_string.character('c')), "abc"),
                     'c');
             }],
             ["P fail Then",
@@ -17,8 +17,8 @@ define(['parse/parse', 'parse/parse_lang'], function(parse, parse_lang){
                 var a = parse.either(
                     parse_lang.then(
                         parse.never(),
-                        parse.character('b')),
-                    parse.character('a'));
+                        parse_string.character('b')),
+                    parse_string.character('a'));
                 
                 assert.deepEqual(parse.run(a, "ab"), 'a');
             }],
@@ -26,9 +26,9 @@ define(['parse/parse', 'parse/parse_lang'], function(parse, parse_lang){
             function(){
                 var a = parse.either(
                     parse.attempt(parse_lang.then(
-                        parse.character('a'),
+                        parse_string.character('a'),
                         parse.never('b'))),
-                    parse.character('a'));
+                    parse_string.character('a'));
                 
                 assert.deepEqual(parse.run(a, "ab"), 'a');
             }],